+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "telcom-parser",
     "installer",
     "installer-gui/src-tauri",
+    "wifi",
 ]
 # at least for now, let's keep installer-gui out of the list of default
 # packages. installer-gui is still experimental and requires many new packages
@@ -19,5 +20,6 @@ default-members = [
     "rootshell",
     "telcom-parser",
     "installer",
+    "wifi",
 ]
 resolver = "2"
