+++ installer/Cargo.toml
@@ -16,6 +16,7 @@ path = "src/main.rs"
 
 [dependencies]
 rayhunter = { path = "../lib" }
+rayhunter-wifi = { path = "../wifi" }
 aes = "0.8.4"
 anyhow = "1.0.98"
 axum = { version = "0.8.3", features = ["http1", "tokio"], default-features = false }
